+++ tests/unit/common/vitals/interaction-to-next-paint.test.js
@@ -12,7 +12,8 @@ const inpAttribution = {
   nextPaintTime: 200,
   processingDuration: 0,
   presentationDelay: 0,
-  loadState: 'complete'
+  loadState: 'complete',
+  interactionTargetElement: document.createElement('div')
 }
 const getFreshINPImport = async (codeToRun) => {
   jest.doMock('web-vitals/attribution', () => ({
@@ -24,8 +25,10 @@ const getFreshINPImport = async (codeToRun) => {
 
 describe('inp', () => {
   test('reports inp from web-vitals', (done) => {
-    getFreshINPImport(metric => metric.subscribe(({ value, attrs }) => {
+    getFreshINPImport(metric => metric.subscribe(({ value, element, attrs }) => {
       expect(value).toEqual(8)
+      expect(element).toEqual(inpAttribution.interactionTargetElement)
+      expect(element instanceof HTMLElement).toBe(true)
       expect(attrs).toStrictEqual({
         eventTarget: inpAttribution.interactionTarget,
         eventTime: inpAttribution.interactionTime,